+++ R/laplace.R
@@ -80,7 +80,7 @@
 #'
 #' }
 #'
-laplaceControl <- function(sigdig=3, ..., nAGQ=1) {
+laplaceControl <- function(sigdig=4, ..., nAGQ=1) {
   # interaction forces the calculation of the hessian, which is needed
   # for the adaptive Gaussian quadrature
   .control <- foceiControl(sigdig=sigdig,